Remove vestigial rule
I talked to @jgwhite who told me that this rule was a spike
that @fivetanley subsequently built up into the existing
rules, so it’s save to remove now! If anyone needs something
like this, it’s covered by the a11y-audit-called rule, which
I will add documentation for separately.